requirements.txt
discord.py==2.3.2
requests==2.31.0
yt-dlp==2023.12.30
youtube-search-python==1.6.6